chore: Separated actor from agent workflow orchestrator. broken: Agent workflow stops executing after the first handoff.
