+++ src/dqa/mcp/primary.py
@@ -3,7 +3,7 @@ import signal
 import sys
 
 from fastmcp import FastMCP, Client
-from dqa import ParsedEnvVars
+from dqa import EnvVars
 
 from dqa.mcp.ollama import app as ollama_mcp
 from dqa.mcp.datetime import app as datetime_mcp
@@ -34,9 +34,11 @@ def server():
     server.mount(yfmcp_mcp, prefix="yfmcp", as_proxy=True)
     logger.info("Mounted YFMCP.")
 
-    if ParsedEnvVars().API_KEY_ALPHAVANTAGE:
+    if EnvVars.API_KEY_ALPHAVANTAGE:
         # See: https://github.com/alphavantage/alpha_vantage_mcp
-        alphavantage_remote_url = f"https://mcp.alphavantage.co/mcp?apikey={ParsedEnvVars().API_KEY_ALPHAVANTAGE}"
+        alphavantage_remote_url = (
+            f"https://mcp.alphavantage.co/mcp?apikey={EnvVars.API_KEY_ALPHAVANTAGE}"
+        )
         alphavantage_remote_proxy = FastMCP.as_proxy(Client(alphavantage_remote_url))
         server.mount(alphavantage_remote_proxy, prefix="alphavantage")
         logger.info("Mounted AlphaVantage remote MCP.")
@@ -48,10 +50,10 @@ def server():
             "No AlphaVantage API key found in environment variable 'API_KEY_ALPHAVANTAGE'. Skipping mounting AlphaVantage remote MCP."
         )
 
-    if ParsedEnvVars().API_KEY_TAVILY:
+    if EnvVars.API_KEY_TAVILY:
         # See: https://github.com/alphavantage/alpha_vantage_mcp
         tavily_remote_url = (
-            f"https://mcp.tavily.com/mcp/?tavilyApiKey={ParsedEnvVars().API_KEY_TAVILY}"
+            f"https://mcp.tavily.com/mcp/?tavilyApiKey={EnvVars.API_KEY_TAVILY}"
         )
         tavily_remote_proxy = FastMCP.as_proxy(Client(tavily_remote_url))
         # Prefix is not necessary because Tavily tools are prefixed with "tavily_" already.
@@ -65,7 +67,7 @@ def server():
             "No AlphaVantage API key found in environment variable 'TAVILY_API_KEY'. Skipping mounting Tavily remote MCP."
         )
 
-    if ParsedEnvVars().API_KEY_OLLAMA:
+    if EnvVars.API_KEY_OLLAMA:
         server.mount(ollama_mcp(), prefix="ollama")
         logger.info("Mounted Ollama MCP.")
     else:
@@ -87,7 +89,7 @@ def main():  # pragma: no cover
 
     signal.signal(signal.SIGINT, sigint_handler)
 
-    transport_type = ParsedEnvVars().DQA_MCP_SERVER_TRANSPORT
+    transport_type = EnvVars.DQA_MCP_SERVER_TRANSPORT
 
     app = server()
 
@@ -99,8 +101,8 @@ def main():  # pragma: no cover
     else:
         app.run(
             transport=transport_type,
-            host=ParsedEnvVars().MCP_SERVER_HOST,
-            port=ParsedEnvVars().MCP_SERVER_PORT,
+            host=EnvVars.MCP_SERVER_HOST,
+            port=EnvVars.MCP_SERVER_PORT,
             uvicorn_config={
                 "timeout_graceful_shutdown": 5,  # seconds
             },
